Push seeker relic textures into SeekerMod
