+++ src/main/java/fruitymod/relics/Telescope.java
@@ -5,14 +5,14 @@ import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.relics.AbstractRelic;
 
 import basemod.abstracts.CustomRelic;
-import fruitymod.FruityMod;
+import fruitymod.SeekerMod;
 import fruitymod.actions.unique.TelescopeAction;
 
 public class Telescope extends CustomRelic {
 	private static final String ID = "Telescope";
 
 	public Telescope() {
-		super(ID, FruityMod.getTelescopeTexture(),
+		super(ID, SeekerMod.getTelescopeTexture(),
 				RelicTier.RARE, LandingSound.MAGICAL);
 	}
